Pass credentials to DB through env file
It's worth considering to just use the environment variables that PostgresSQL relies on in our Django code, and so avoid having two identical environment variables, but on the other hand, that might make our code too tightly coupled with Postgres. I'm not sure which one is better. I'm open to a good argument on why I should go the other way on this one.