Add make messages method: print_bold
Of course the "oh, this is small enough that we don't need to do it properly" script got so big that we totally need to do it properly now. That's always what happens. What ever made you think this time would be the exception? It's never the exception, not ever.